// common/randomizer/enums/PrimeItemId.ts
export enum PrimeItemId {
  MISSILE_EXPANSION = 0x04,
  ENERGY_TANK = 0x18,
  POWER_BOMB_EXPANSION = 0x07,
  MISSILE_LAUNCHER = 0x04,
  WAVE_BEAM = 0x02,
  ICE_BEAM = 0x01,
  PLASMA_BEAM = 0x03,
  CHARGE_BEAM = 0x0a,
  SPACE_JUMP_BOOTS = 0x0f,
  SUPER_MISSILE = 0x0b,
  WAVEBUSTER = 0x1c,
  ICE_SPREADER = 0x0e,
  FLAMETHROWER = 0x08,
  GRAPPLE_BEAM = 0x0c,
  MORPH_BALL = 0x10,
  BOOST_BALL = 0x12,
  SPIDER_BALL = 0x13,
  MORPH_BALL_BOMB = 0x06,
  POWER_BOMB = 0x07,
  VARIA_SUIT = 0x16,
  GRAVITY_SUIT = 0x15,
  PHAZON_SUIT = 0x17,
  THERMAL_VISOR = 0x09,
  XRAY_VISOR = 0x0d,
  ARTIFACT_OF_TRUTH = 0x1d,
  ARTIFACT_OF_STRENGTH = 0x1e,
  ARTIFACT_OF_ELDER = 0x1f,
  ARTIFACT_OF_WILD = 0x20,
  ARTIFACT_OF_LIFEGIVER = 0x21,
  ARTIFACT_OF_WARRIOR = 0x22,
  ARTIFACT_OF_CHOZO = 0x23,
  ARTIFACT_OF_NATURE = 0x24,
  ARTIFACT_OF_SUN = 0x25,
  ARTIFACT_OF_WORLD = 0x26,
  ARTIFACT_OF_SPIRIT = 0x27,
  ARTIFACT_OF_NEWBORN = 0x28
}